BaseTools/FMMT: FMMT python tool with view function
The FMMT python tool is under development, which will have the
Fv/FFs-based 'View'&'Add'&'Delete'&'Replace' operation function.

This patch is the first development stage of FMMT python tool,
which realize the 'View' function. When giving a Fd/Fv binary file,
the 'View' function can show the binary structure of the file with
an output log file.

Signed-off-by: Yuwei Chen <yuwei.chen@intel.com>
Cc: Bob Feng <bob.c.feng@intel.com>