BaseTools/FMMT:Section compress function & ctypes Header
The FMMT python tool is under development, which will have the
Fv/FFs-based 'View'&'Add'&'Delete'&'Replace' operation function.

This patch supports the compression function of section and changes
the FV/FFS/SECTION Header to ctypes which is consistent with the
definition in the spec and easy for decoding and encoding.

Signed-off-by: Yuwei Chen <yuwei.chen@intel.com>
Cc: Bob Feng <bob.c.feng@intel.com>